Fix the hasType() AST matcher to not assert when the QualType is invalid.
There's not a particularly good way to test this with the AST matchers unit tests because the only way to get an invalid type (that I can devise) involves creating parse errors, which the test harness always treats as a failure. Instead, a clang-tidy test case will be added in a follow-up commit based on the original bug report.

git-svn-id: https://llvm.org/svn/llvm-project/cfe/trunk@326604 91177308-0d34-0410-b5e6-96231b3b80d8